+++ packages/plugin-oauth-github/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @swr-login/plugin-oauth-github
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @swr-login/core@0.4.0
+
 ## 0.2.1
 
 ### Patch Changes